set role assignments via api;
