+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/api/CFTLib.java
@@ -1,7 +1,14 @@
 package uk.gov.hmcts.rse.ccd.lib.api;
 
+import java.nio.charset.Charset;
+import java.sql.Connection;
 import java.util.List;
+import javax.sql.DataSource;
+import lombok.SneakyThrows;
+import org.apache.commons.io.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.DefaultResourceLoader;
+import org.springframework.core.io.ResourceLoader;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.ccd.definition.store.repository.SecurityClassification;
 import uk.gov.hmcts.ccd.definition.store.repository.model.UserRole;
@@ -11,6 +18,9 @@ import uk.gov.hmcts.ccd.endpoint.userprofile.UserProfileEndpoint;
 
 @Component
 public class CFTLib {
+  @Autowired
+  private DataSource data;
+
   @Autowired
   UserRoleController roleController;
 
@@ -33,6 +43,23 @@ public class CFTLib {
       r.setRole(role);
       r.setSecurityClassification(SecurityClassification.PUBLIC);
       roleController.userRolePut(r);
+    }
+  }
+
+  @SneakyThrows
+  public void configureRoleAssignments(String json){
+    try (Connection c = data.getConnection()) {
+      // To use the uuid generation function.
+      c.createStatement().execute(
+          "create extension pgcrypto"
+      );
+
+      ResourceLoader resourceLoader = new DefaultResourceLoader();
+      // Provided by the consuming application.
+      var sql = IOUtils.toString(resourceLoader.getResource("classpath:rse/cftlib-populate-am.sql").getInputStream(), Charset.defaultCharset());
+      var p = c.prepareStatement(sql);
+      p.setString(1, json);
+      p.executeQuery();
     }
   }
 }
